big changes
- icon normalization
- energy is now analogue
- every reagent now has a calorie count, and is used to determine how much energy is gained.
- many many new food items and reactions.

+++ Assets/KoboldKare/Scripts/EnergyBarDisplayUI.cs
@@ -24,7 +24,7 @@ public class EnergyBarDisplayUI : MonoBehaviour {
         running = false;
     }
 
-    void OnEnergyChanged(int energy, int maxEnergy) {
+    void OnEnergyChanged(float energy, float maxEnergy) {
         targetWidth = energy * energyToPixel;
         targetBackgroundWidth = maxEnergy * energyToPixel;
         if (!running) {
